Exercise 4.21 from part 4: Restriction on delete method by token

+++ part4/blog/backend/controllers/blogsRouter.js
@@ -38,7 +38,18 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog.user.toString() === user.id.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+  }else{
+    return response.status(401).json({error: 'you cant delete this blog'})
+  }
   response.status(204).end()
 })
